Update ascii.properties with ASCII 32 [Space]
Not having a value defined for Ascii Character 32 causes an issue allowing those running the pack to type a greater number of spaces in a sign than those who do not. This then causes parts of a typed sign to not render for default / alternate resource packs. Giving ASCII Character 32, the space character, a width of 4 seems to mimic the spacing of Minecraft's Default ASCII 32 Character, allowing for identical rendering of signs between resource packs.